[CDG-47][CDG-59] - Separate services in folders and update its imports
